firstproject/requirement.txt
django==4.0.1
djangorestframework==3.13.1
mysqlclient==2.01
django-mysql==3.9
django-cors-headers==3.11.0
pika==1.1.0